ESP32 Support - EEPROM/Flash
FlashStorage isn't compatible with ESP32 boards (and probably not anytime soon), so I needed to use the arduino-esp32 built in library for handling the flash and tailor the calls for it using the Preferences lib. It could also be in the normal Arduino core but I haven't looked into it.

- Added header defines to handle ESP32 boards
- Added ESP32 specific get and set calls for the flash memory
- In `EEPROM_METADATA_T`, specified a definition of the metadata loader for ESP32 boards using the Preferences library